requirements.txt
absl-py==1.0.0
Adv2==1.2.0
appnope==0.1.2
asgiref==3.4.1
astroid==2.7.2
astropy==4.2.1
astroquery==0.4.3
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
CacheControl==0.12.6
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.2
click==8.0.3
colorama==0.4.4
cycler==0.10.0
debugpy==1.3.0
decorator==4.4.2
Django==3.2.8
django-cors-headers==3.10.0
django-recaptcha==2.0.6
django-resized==0.3.11
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
docopt==0.6.2
firebase-admin==5.0.3
flatbuffers==2.0
future==0.18.2
gast==0.4.0
google-api-core==2.1.1
google-api-python-client==2.26.1
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.1.0
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
h5py==3.6.0
html5lib==1.1
httplib2==0.20.1
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.4
importlib-metadata==4.6.1
imutils==0.5.4
inflect==5.3.0
iniconfig==1.1.1
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
isort==5.9.3
jedi==0.18.0
joblib==1.1.0
JsonForm==0.0.2
jsonschema==3.2.0
JsonSir==0.0.2
jupyter-client==6.1.12
jupyter-core==4.7.1
keras==2.7.0
Keras-Preprocessing==1.1.2
keyring==23.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
libclang==12.0.0
llvmlite==0.36.0
mando==0.6.4
Markdown==3.3.6
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mimeparse==0.1.3
more-itertools==8.8.0
moviepy==1.0.3
msgpack==1.0.2
networkx==2.5.1
nltk==3.6.5
num2words==0.5.10
numba==0.53.1
numpy==1.21.0
oauthlib==3.1.1
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.4
parso==0.8.2
pdfrw==0.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
platformdirs==2.2.0
pluggy==0.13.1
proglog==0.1.9
prompt-toolkit==3.0.19
proto-plus==1.19.5
protobuf==3.18.1
psycopg2-binary==2.9.2
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyerfa==2.0.0
Pygments==2.9.0
PyInvoice==0.1.7
PyJWT==2.2.0
pylint==2.10.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pymovie==2.9.1
pyparsing==2.4.7
PyQRCode==1.2.1
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqtgraph==0.12.2
pyrsistent==0.18.0
pytesseract==0.3.8
pytest==6.2.4
pytest-django==4.4.0
python-dateutil==2.8.1
python-decouple==3.5
Python-EasyConfig==0.1.7
pytz==2021.3
pyvo==1.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
radon==5.1.0
regex==2021.10.23
reportlab==3.6.2
requests==2.26.0
requests-oauthlib==1.3.0
Resource==0.2.1
rsa==4.7.2
scikit-image==0.18.2
scipy==1.7.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
tifffile==2021.7.2
toml==0.10.2
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.5.0
gunicorn==20.1.0
